Fixed serialize party
